requirements.txt
python-dotenv
pandas 
numpy
pymongo 
certifi
pymongo[srv]
scikit-learn
mlflow
pyaml
dagshub
fastapi
uvicorn
python-multipart
dill